feature specialization fixes build errors + clippy

+++ field/src/field.rs
@@ -156,7 +156,7 @@ default impl<F: PrimeField64> Field for F {
     default type Packing = F;
     default fn try_inverse(&self) -> Option<Self> {
         crate::inverse_u::<u64, i64>(self.as_canonical_u64(), Self::ORDER_U64)
-            .map(|x| Self::from_canonical_u64(x as u64))
+            .map(|x| Self::from_canonical_u64(x))
     }
 }
 
@@ -164,7 +164,7 @@ default impl<F: PrimeField32> Field for F {
     default type Packing = F;
     default fn try_inverse(&self) -> Option<Self> {
         crate::inverse_u::<u32, i32>(self.as_canonical_u32(), Self::ORDER_U32)
-            .map(|x| Self::from_canonical_u32(x as u32))
+            .map(|x| Self::from_canonical_u32(x))
     }
 }
